Add support for HTTP/2 ALPN negotiation.
Motivation

The most common mechanism for negotiating HTTP/2 is to use ALPN.
We should support that.

Modifications:

Added a `ChannelPipeline` extension to configure a pipeline for
ALPN-based upgrade.

Added a property that indicates the ALPN tokens for HTTP/2 and
HTTP/1.1 use.

Result:

Should be possible to configure a channel pipeline for HTTP/2 work.