Clarify cleanup docs for stream channel creation
Motivation:

The documentation for creating a new stream channel didn't indicate
whether the user had to do any cleanup on the multiplexer once a stream
was no longer in use.

Modifications:

Added a note to the documentation indicating that no additional action
is required by the user beyond closing the stream.

Result:

The documentation is less ambiguous.