Mark entire StringRing as readable.
Motivation:

We use a Bytebuffer as backing storage for a string ringbuffer to avoid
needing to reinvent that data structure ourselves. This used to work great
but NIO 2 recently changed the get* functions to trap if they accessed
memory outside the reader/writer indices.

We were never moving the writer index, instead treating the entire buffer
as valid. This led to preconditions, and sadness.

In an ideal world we'd only mark the bytes that are actually active as
readable to ensure that we have the defensive traps from bytebuffer. In
practice, that's a sufficiently complex thing to do that we're better
served just trusting our code and taking off the guard rails.

In future we should go back to seeing whether a CircularBuffer of strings
will work better, now that Strings are fast. I'll make that Johannes'
problem.

Modifications:

- Mark the entire backing buffer as readable.

Result:

Fewer crashes