Avoid double-counting of flow controlled length of synthesised data frames.
Motivation:

When synthesising data frames we try to emit the flow controlled length of the entire real
data frame on the first synthetic frame. This allows us to correctly police the window management
of that frame, while still synthesising frames.

However, @glbrntt spotted that we were incorrectly repeatedly counting the flow controlled length
of a partial frame. A patch that had been previously added to ensure we only counted it once was
not working correctly.

Modifications:

- Adjusted the location of the call to flowControlledLength.

Result:

Correct accounting of flow controlled length