Correctly persist frame flush status.
Motivation:

We recently spotted that in an edge case it was possible for the
outbound frame buffer to incorrectly believe that a flushed frame was in
fact not flushed, and to refuse to give it up until it had encountered a
flush point. This could happen if the frame was flushed but buffered in
the concurrent streams buffer, and was then moved to the flow control
buffer but could not be emitted.

Modifications:

Correctly marked such frames as flushed.

Result:

We won't wedge frames in the buffer.