Close stream channels when the parent channel becomes inactive (#131)
Motivation:

Stream channels don't get closed when the parent channel closes. This
can lead to some unexpected behaviour for users.

Modifications:

Call receiveStreamClosed on stream channels in the multiplexer when
reciving channelInactive. Added a test to verify the stream is closed.

Result:

Stream channels will be closed when the parent channel is closed.