Fix partial table lookups. (#168)
Motivation:

In many cases it's necessary to be able to do HPACK table lookups to
find only a matching header field name, when it isn't going to be
possible to use a fully indexed header. For example, when using
nonIndexable header fields you cannot use an indexed field value, so
there is no point looking at all the header field values anyway.

While we technically supported this, there were two wrinkles. The first
was that if you attempted to search for a header field without any care
for its name, we'd actually only search the static table, missing
anything added into the dynamic table, which is pretty stupid.

The second issue was that nonIndexable header fields did extra
unnecessary work. This is because we asked the header table to search
for an exact match, but then disregarded whether or not one was found.
That meant a bunch of wasted computation.

Modifications:

- Fixed the table lookup to correctly search both dynamic and static
    tables for partial matches.
- Correctly indicated that nonIndexable headers shouldn't care about
    matching the header value.

Result:

Better handling of non-indexable headers.