tolerate multiple RST_STREAMs for same stream (#171)
Motivation:

The HTTP/2 RFC does not explicitly forbid sending more than one
RST_STREAM for the same stream. Given that we have seen this behaviour
in the wild we are going to ignore any additional RST_STREAMs.

A further indicative that this is the right call is that the RFC
explicitly mentions that sending RST_STREAMs in the idle state is not
acceptable.

Modifications:

Ignore RST_STREAM frames for streams in the closed state.

Result:

More compatibility.