Update the test generation script to allow for deprecated tests
Motivation:

To test deprecated functionality without a bunch of warnings, tests must
be marked as deprecated. This also needs to be bubbled up to anywhere
calling that test, such as in the generated linux test manifests.

Our script to generate the manifests currently doesn't support this and
as part of #214 we'll need to deprecate a few things while keeping their
tests.

Modifications:

- Pull in the latest version of the `generate_linux_tests.rb` script
  from SwiftNIO.
- Re-run the script, since it includes some formatting changes too.

Result:

- When we deprecate functionality, we can deprecate the tests without
  the generate manifests emitting warnings.