Shrink NIOHTTP2WindowUpdatedEvent. (#239)
Motivation:

NIOHTTP2WindowUpdatedEvent was larger than 24 bytes, forcing heap
allocations whenever we wanted to put it into an existential. Given that
the range of the two optional integers stored on the type are both
explicitly documented to be the Int32 range, this type can happily just
contain Int32s.

Modifications:

- Changed the backing storage from Int? to Int32?.

Result:

NIOHTTP2WindowUpdatedEvent will be smaller.